+++ CHANGES.rst
@@ -12,6 +12,8 @@ Features
 
 - Add a cron job which does a daily backup of the database.
 
+- Allow to edit the labels of pre-defined fields.
+
 Changes in testing
 ------------------